requirements.txt
flask==2.0.1
requests==2.26.0
python-binance==1.0.25
pandas==2.3.0
numpy==1.21.2
schedule==1.1.0
